* reduce font-size texts * rename buttonStyles stylesheet * share modal title and subtitle styles
