+++ src/client/src/components/SubscriptionSelection/index.tsx
@@ -1,7 +1,10 @@
 import * as React from "react";
 import { injectIntl, InjectedIntlProps } from "react-intl";
+
+import modalStyles from "../../css/modal.module.css";
 import styles from "./styles.module.css";
 import messages from "./messages";
+
 import Dropdown from "../Dropdown";
 import { useSelector } from "react-redux";
 import { getDropdownSubscriptionsSelector } from "../../store/config/azure/selector";
@@ -37,11 +40,11 @@ const SubscriptionSelection = (props: Props) => {
 
   return (
     <div className={styles.container}>
-      <div className={styles.header}>
-        <div className={styles.title}>{formatMessage(messages.title)}</div>
+      <div className={modalStyles.header}>
+        <div className={modalStyles.title}>{formatMessage(messages.title)}</div>
         <a href={AZURE_LINKS.CREATE_NEW_SUBSCRIPTION}>{formatMessage(messages.newSubscriptionLink)}</a>
       </div>
-      <div className={styles.subtitle}>{formatMessage(messages.subtitle)}</div>
+      <div className={modalStyles.subtitle}>{formatMessage(messages.subtitle)}</div>
       <Dropdown
         ariaLabel={formatMessage(messages.ariaDropdownLabel)}
         options={subscriptions}
