+++ src/client/src/css/modal.module.css
@@ -1,0 +1,17 @@
+.header {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 3px;
+  vertical-align: baseline;
+}
+
+.title {
+  font-weight: 500;
+  font-size: 1.2em;
+}
+
+.subtitle {
+  margin-bottom: 1em;
+  font-size: 90%;
+  opacity: 0.8;
+}
