+++ src/client/src/modals/AppServiceModal/AppServicePlanInfo/index.tsx
@@ -1,7 +1,10 @@
 import * as React from "react";
 import { injectIntl, InjectedIntlProps } from "react-intl";
+
+import modalStyles from "../../../css/modal.module.css";
 import styles from "./styles.module.css";
 import messages from "./messages";
+
 import { AZURE_LINKS } from "../../../utils/constants/azure";
 import { AppState } from "../../../store/combineReducers";
 import { useSelector } from "react-redux";
@@ -25,13 +28,13 @@ const AppServicePlanInfo = (props: Props) => {
 
   return (
     <div className={styles.container}>
-      <div className={styles.title}>{formatMessage(messages.title)}</div>
-      <div>
+      <div className={modalStyles.title}>{formatMessage(messages.title)}</div>
+      <div className={modalStyles.subtitle}>
         {isMicrosoftLearnSubscription()
           ? formatMessage(messages.freeTierMessage)
           : formatMessage(messages.basicTierMessage)}
       </div>
-      <a target={"_blank"} rel="noreferrer noopener" href={AZURE_LINKS.APP_SERVICE_PLAN}>
+      <a className={modalStyles.subtitle} target={"_blank"} rel="noreferrer noopener" href={AZURE_LINKS.APP_SERVICE_PLAN}>
         {formatMessage(messages.learnMore)}
       </a>
     </div>
