+++ src/client/src/modals/AppServiceModal/RuntimeStackInfo/index.tsx
@@ -1,9 +1,12 @@
 import * as React from "react";
 import { useSelector } from "react-redux";
 import { injectIntl, InjectedIntlProps } from "react-intl";
+
+import modalStyles from "../../../css/modal.module.css";
 import styles from "./styles.module.css";
-import { AppState } from "../../../store/combineReducers";
 import messages from "./messages";
+
+import { AppState } from "../../../store/combineReducers";
 import { getRuntimeStackSelector } from "../../../store/userSelection/app/selector";
 
 const RuntimeStackInfo = ({ intl }: InjectedIntlProps) => {
@@ -11,7 +14,7 @@ const RuntimeStackInfo = ({ intl }: InjectedIntlProps) => {
   const runtimeStack = useSelector((state: AppState) => getRuntimeStackSelector(state));
   return (
     <div className={styles.container}>
-      <div className={styles.title}>{formatMessage(messages.title)}</div>
+      <div className={modalStyles.title}>{formatMessage(messages.title)}</div>
       {formatMessage(messages.runtimeStack, { runtimeStack })}
     </div>
   );
