+++ src/client/src/modals/AppServiceModal/RuntimeStackInfo/styles.module.css
@@ -1,9 +1,3 @@
 .container {
   margin-bottom: 35px;
 }
-
-.title {
-  font-weight: 500;
-  font-size: 1.2em;
-  margin-bottom: 3px;
-}
