+++ src/client/src/pages/AddPagesPage/PageCard/index.tsx
@@ -19,7 +19,7 @@ import classNames from "classnames";
 import styles from "./styles.module.css";
 import cardStyles from "../../cardStyles.module.css";
 import pageStyles from "../../cardStyles.module.css";
-import buttonStyles from "../../../css/buttonStyles.module.css";
+import buttonStyles from "../../../css/button.module.css";
 
 type Props = IProps & IStateProps & InjectedIntlProps;
