+++ src/client/src/pages/AddPagesPage/PageCard/styles.module.css
@@ -11,7 +11,6 @@
 /* below similar but not exactly same styles */
 /* TODO have a look at this as this box should have 3 rows in each page card here  */
 .description {
-  font-size: 1.2rem;
   line-height: 1.5;
   margin-top: 1rem;
   height: 9rem;
