+++ src/client/src/pages/ReviewAndGeneratePage/index.tsx
@@ -4,7 +4,7 @@ import { useDispatch } from "react-redux";
 import * as ModalActions from "../../store/navigation/modals/action";
 
 import styles from "./styles.module.css";
-import buttonStyles from "../../css/buttonStyles.module.css";
+import buttonStyles from "../../css/button.module.css";
 
 import { InjectedIntlProps, injectIntl } from "react-intl";
 import messages from "./messages";
