+++ src/client/src/pages/ReviewAndGeneratePage/styles.module.css
@@ -9,7 +9,6 @@
 }
 
 .selectionContainer > p {
-  font-size: 1.2em;
   max-width: 350px;
 }
