+++ src/client/src/pages/SelectFrameworksPage/FrameworkCard/styles.module.css
@@ -5,7 +5,6 @@
 
 .latestVersion {
   margin-top: 0.5rem;
-  font-size: 1.2rem;
   line-height: 1.5;
   padding-left: 0.5rem;
 }
